+++ src/Taxonomies/Term.php
@@ -7,7 +7,9 @@ use Statamic\Data\ExistsAsFile;
 use Statamic\Data\SyncsOriginalState;
 use Statamic\Events\TermBlueprintFound;
 use Statamic\Events\TermCreated;
+use Statamic\Events\TermCreating;
 use Statamic\Events\TermDeleted;
+use Statamic\Events\TermDeleting;
 use Statamic\Events\TermSaved;
 use Statamic\Events\TermSaving;
 use Statamic\Facades;
@@ -211,6 +213,10 @@ class Term implements TermContract
         $this->afterSaveCallbacks = [];
 
         if ($withEvents) {
+            if ($isNew && TermCreating::dispatch($this) === false) {
+                return false;
+            }
+
             if (TermSaving::dispatch($this) === false) {
                 return false;
             }
@@ -237,6 +243,10 @@ class Term implements TermContract
 
     public function delete()
     {
+        if (TermDeleting::dispatch($this) === false) {
+            return false;
+        }
+
         Facades\Term::delete($this);
 
         TermDeleted::dispatch($this);
